Add slug field to Page model and update createPage method
The Page model now includes a new field called 'slug'. This field has been added to the fillable array in the model to allow it to be mass assignable. Additionally, the 'createPage' method in the PageService class has been updated to generate a unique slug using the Str::uuid() method before creating a new page.